Reshuffle levity polymorphism checks.
Previously, GHC checked for bad levity polymorphism to the
left of all arrows in data constructors. This was wrong, as
reported in #??? (where an example is also shown). The solution
is to check each individual argument for bad levity polymorphism.
Thus the check has been moved from TcValidity to TcTyClsDecls.

A similar situation exists with pattern synonyms, also fixed here.

This patch also nabs #12819 while I was in town.

Test cases: ???